+++ src/scrapeghost/__init__.py
@@ -1,7 +1,5 @@
 # ruff: noqa
 from .scrapers import (
-    BadStop,
-    InvalidJSON,
     PaginatedSchemaScraper,
     SchemaScraper,
 )