+++ tests/test_nudge_live.py
@@ -1,10 +1,10 @@
-from scrapeghost import SchemaScraper, InvalidJSON
+from scrapeghost import SchemaScraper
+from scrapeghost.postprocessors import JSONPostprocessor
 
 
 def test_bad_json_nudge():
     # single quotes, trailing commas
-    e = InvalidJSON("Invalid JSON", "{'name': 'phil', }")
-    schema = {"name": "string"}
-    scraper = SchemaScraper(schema)
-    result = scraper.nudge(e)
-    assert result.message.content == '{"name": "phil"}'
+    bad_json = "{'name': 'phil', }"
+    jpp = JSONPostprocessor(nudge=True)
+    repaired = jpp(scraper=SchemaScraper({"name": "string"}), data=bad_json)
+    assert repaired == {"name": "phil"}